backend-django/requirements.txt
# Django Core
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Redis & Celery
redis==5.0.3
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1

# API Documentation
drf-spectacular==0.27.2

# Authentication & Security
PyJWT==2.8.0
python-decouple==3.8

# Storage
boto3==1.34.69
minio==7.2.5

# Monitoring
prometheus-client==0.20.0

# Logging
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
requests==2.31.0
pytz==2024.1

# GIS
django-geojson==4.1.0

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-mock==3.14.0
factory-boy==3.3.0
faker==24.4.0

# Code Quality
black==24.3.0
flake8==7.0.0
isort==5.13.2
mypy==1.9.0
django-stubs==4.2.7

# ASGI Server
gunicorn==22.0.0
uvicorn[standard]==0.29.0

# Others
Pillow==10.3.0
python-magic==0.4.27
